CardConnect: Remove domain validation
CardConnect no longer requires ports to be included in their valid
domains, so this change removes that requirement, allowing customers to
use CardConnect domains without specifying a port.

[CE-209](https://spreedly.atlassian.net/browse/CE-209)

Unit:
22 tests, 92 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

Remote:
25 tests, 58 assertions, 2 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
92% passed

Two failing remote tests:
 - `test_successful_refund`
 - `test_partial_refund`
Both of these tests are failing with the message "Txn not settled" and
are failing on master as well. The last commit to CardConnect (Oct 2019)
shows fully passing remote tests, so something appears to have changed
in the gateway integration since then. These new failures seem to be
unrelated to the changes made in this commit.

All unit tests:
4414 tests, 71334 assertions, 0 failures, 0 errors, 0 pendings, 2
omissions, 0 notifications
100% passed